fix: 2026-09 audit — crashes, remote parsers, lifecycle races, ~215 KB internal SRAM
Eight-agent code audit of the firmware (no hardware); everything below was
verified by reading the code paths and the build is green. Not yet flashed.

Crash / memory safety
- nv_keydeck: task ran NVS reads on a PSRAM stack (abort) -> internal 8 KB.
- nv_anima: nucleo_arb_init() was never called, so every online TLS call was
  refused as "busy" (the whole online tier was dead); init with the engine,
  fail-open before init. Engine init is now one-shot under the spine gate
  (both surfaces re-ran it per query and unloaded L1 under each other).
  L1 clarify picks re-open the index file the offsets belong to (NULL fseek
  panic on "1"/"2"). a_subst_regs snprintf overflow, facet result left
  uninitialised, profile RECALL word boundary, a_from_base wrap, fold_calc
  truncation, prime trial-division bound, learn/online/conv stores no longer
  destroy both copies on a rename failure (commit_tmp), teacher.json read by
  size (1.5 KB stack reads dropped keys), 256-char provider keys, breaker not
  poisoned by local bails, persona text updated for this hardware, memory
  facts deduplicated, localtime_r.
- nv_vplayer: MP4 AVCC length and AVI chunk strides computed in 64-bit with
  monotonic-progress checks (remote infinite loop / OOB memcpy via
  /api/video/play), stsz/stts bounds, seek re-open double free, pause pacer
  re-anchor, MPEG-1 pcm_end only after a successful begin (FreeRTOS mutex
  assert), removal-safe file session, AAC audio task stack 20 KB.
- nv_media: WAV header window 4 KB, byterate 0, unsupported layouts rejected,
  NULL decoder close, EOT cleared on stop; MP3 decoder state/scratch/pcm
  allocated per playback (nv_mp3dec_reset/release, minimp3 scratch pointer).
- nv_web: path guard rejects '\' and '//', settings.nvb never served, request
  bodies in PSRAM with an error reply on short reads, oversized WS frames close
  the session, intent/arg JSON-escaped, no inline ANIMA fallback on the 8 KB
  httpd stack, /api/ui/home posted to the LVGL thread.
- nv_wasm: guest ints clamped in gfx_line/gfx_text/gfx_image (native loops the
  watchdog cannot stop), image cache byte budget, nv.save cannot overwrite the
  manifest/module, 8 KB manifests, canvas size clamped in read_manifest,
  recursive uninstall, abort fallback ordering, module/worker buffers in PSRAM.
- nv_ota / nv_appstore: response collectors truncate+flag instead of dropping
  chunks, PSRAM manifest buffers, TLS-sized task stacks, download capped at the
  OTA slot size, mark-valid runs on its own task, updates wait out the 60 s
  survival gate, install refused while the app runs.
- nv_hal: nv_backup export moved off the esp_timer task (it froze the LVGL
  tick on every app launch), nv_sd drains outside its lock and nv_sd_info uses
  a file session (max_files 16), nv_wifi handles esp_wifi_connect() errors
  and a full command queue, netif init idempotent, nv_audio SFX/VOICE begin
  with a 400 ms timeout + nv_audio_pcm_owner(), TTS flushes only its own
  stream and uses removal-safe clip files.
- nv_ui: nv_ui_open_app() tears the caller down first (Anima launches were
  no-ops), launcher order/folders remapped when an app is unregistered,
  registry 40 with overflow log, WASM apps skip the in-place rebuild on
  theme/rotation, smart dock excludes WASM tiles by record, order_save writes
  only what changed, nv_ui_go_home_async().
- nv_apps: gallery HW JPEG buffer 16-aligned (every 1080p photo fell back to
  the SW decoder), viewer raster freed/resized, scan-lock retry; notes new-note
  name collision; anima_app voice dispatch waits for the finalized WAV, query
  text travels with the job, launch timer tracked, vault written via
  tmp+rename and never over an unreadable one; apps manager scans the card
  once per open, game-error path restores Back and hides the canvas; video
  config cached (30-60 NVS reads/s), strips restored on teardown; secondscreen
  lock order, partial frame-buffer alloc, touch task check; music/recorder
  chip cache, stale EOT, mic-idle gates; camera folder validation, MP4 path
  recorded frames again, index OOM drops frames; terminal/settings/apps
  snprintf accumulation clamps; files async cancel; sysmon table sized from
  the live task count with a static mutex.
- nv_usb: HID poll interval 1 ms on high speed, report queue no longer one
  behind, GET_REPORT length honoured, frame pool allocation checked.
- nv_camera: MP4 duration math in 64-bit.

Internal SRAM: static .bss 202 KB -> 40 KB (NV_PSRAM_BSS on cold tables and
handler scratch, lazy MP3 decoder), IRAM text 149 KB -> 100 KB, .data
21.5 KB -> 17 KB. docs/BACKLOG.md lists what was deliberately left for later.

+++ components/nv_camera/nv_mp4.c
@@ -136,14 +136,14 @@ long nv_mp4_close(nv_mp4_t *m){
 
     // mvhd
     { long b = box_begin(f, "mvhd"); fullbox(f,0,0);
-      w32(f,0); w32(f,0); w32(f,1000); w32(f, duration*1000/ts);
+      w32(f,0); w32(f,0); w32(f,1000); w32(f, (uint32_t)((uint64_t)duration * 1000 / ts));   /* 32-bit product wrapped after ~47 s at 30 fps */
       w32(f,0x10000); w16(f,0x100); w16(f,0); w32(f,0); w32(f,0);
       wmatrix(f); for(int i=0;i<6;i++) { w32(f,0); } w32(f,2); box_end(f,b); }
 
     // trak
     long trak = box_begin(f, "trak");
     { long b = box_begin(f, "tkhd"); fullbox(f,0,3 /*enabled|in-movie*/);
-      w32(f,0); w32(f,0); w32(f,1); w32(f,0); w32(f, duration*1000/ts);
+      w32(f,0); w32(f,0); w32(f,1); w32(f,0); w32(f, (uint32_t)((uint64_t)duration * 1000 / ts));   /* 32-bit product wrapped after ~47 s at 30 fps */
       w32(f,0); w32(f,0); w16(f,0); w16(f,0); w16(f,0); w16(f,0);
       wmatrix(f); w32(f, (uint32_t)m->w<<16); w32(f, (uint32_t)m->h<<16); box_end(f,b); }
